requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=2.9
psycopg2-binary>=2.9
python-dotenv>=1.0
flake8>=6.0
pytest>=7.0
pytest-django>=4.5
coverage
